+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+- Vehicle blocks extracted to separate table. Vehicle schedules are
+  now a pivot table between `netex_vehicle_blocks` and
+  `netex_vehicle_schedules`.
+
 ## [0.1.1] – 2022-04-06
 
 ### Added